requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
oauthlib==3.3.1
python-dotenv==1.2.1
requests==2.32.5
requests-oauthlib==2.0.0
tweepy==4.16.0
urllib3==2.6.3
